p_pyro/requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
cloudpickle==0.5.2
decorator==4.2.1
enum34==1.1.6
graphviz==0.8.2
ipdb==0.11
ipython==5.5.0
ipython-genutils==0.2.0
networkx==2.1
numpy==1.14.1
observations==0.1.4
pathlib2==2.3.0
pexpect==4.4.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pyro-ppl==0.1.2
PyYAML==3.12
scandir==1.7
scipy==1.0.0
simplegeneric==0.8.1
six==1.11.0
torch==0.3.0.post4
tqdm==4.19.6
traitlets==4.3.2
wcwidth==0.1.7
